Moved all External Dataset Agent and Data Handler code into the ion/agents/data module
